# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Airtime Vending API",
        "description": "API for vending airtime using MTN and GLO through BAP or Shaggo partners",
        "version": "1.0.0"
    },
    "paths": {
        "/vend/recharge/{type}": {
            "get": {
                "tags": [
                    "Airtime"
                ],
                "summary": "Buy airtime using BAP or Shago",
                "description": "Endpoint for purchasing airtime via BAP or Shago services. The 'type' parameter determines which provider to use.",
                "operationId": "3ad6c37440e2d9f2130228006d7112cd",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "The provider type: 'bap' or 'shago'",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "bap",
                                "shago"
                            ],
                            "example": "bap"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "The phone number to recharge (format: 07012345678). For BAP only.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "07065788819"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "The amount of airtime to purchase (in NGN)",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    },
                    {
                        "name": "service_type",
                        "in": "query",
                        "description": "The network provider for BAP (e.g., MTN, GLO)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MTN",
                                "GLO"
                            ],
                            "example": "MTN"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "The phone number to recharge (for Shago only)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "07065788819"
                        }
                    },
                    {
                        "name": "network",
                        "in": "query",
                        "description": "The network provider for Shago (e.g., MTN, GLO)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MTN",
                                "GLO"
                            ],
                            "example": "MTN"
                        }
                    },
                    {
                        "name": "vend_type",
                        "in": "query",
                        "description": "Vend type for Shago service",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "VTU"
                        }
                    },
                    {
                        "name": "serviceCode",
                        "in": "query",
                        "description": "Service code for Shago service",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "QAB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request or missing required parameters"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Airtime",
            "description": "Airtime"
        }
    ]
}